+++ app/src/main-process/squirrel-updater.ts
